clean up directory test stubbed files
