+++ python/cudf/cudf/tests/conftest.py
@@ -4,9 +4,11 @@ import itertools
 import operator
 import os
 import pathlib
+import zoneinfo
 
 import cupy as cp
 import numpy as np
+import pandas as pd
 import pytest
 
 import rmm  # noqa: F401
@@ -173,6 +175,37 @@ def pytest_runtest_makereport(item, call):
     # Set a report attribute for each phase of a call, which can
     # be "setup", "call", "teardown"
     setattr(item, "report", {rep.when: rep})
+
+
+def _get_all_zones():
+    zones = []
+    for zone in zoneinfo.available_timezones():
+        # TODO: pandas 3.0 defaults to zoneinfo,
+        # so all_zone_names can use zoneinfo.available_timezones()
+        try:
+            pd.DatetimeTZDtype("ns", zone)
+        except Exception:
+            continue
+        else:
+            zones.append(zone)
+    return zones
+
+
+# NOTE: _get_all_zones is a very large list; we likely do NOT want to
+# use it for more than a handful of tests
+@pytest.fixture(params=_get_all_zones())
+def all_timezones(request):
+    return request.param
+
+
+@pytest.fixture(
+    params=["America/New_York", "Asia/Tokyo", "CET", "Etc/GMT+1", "UTC"]
+)
+def limited_timezones(request):
+    """
+    Limited set of interesting timezones for testing.
+    """
+    return request.param
 
 
 @pytest.fixture(
