+++ python/cudf/cudf/tests/indexes/datetimeindex/methods/test_tz_convert.py
@@ -2,6 +2,7 @@
 import zoneinfo
 
 import pandas as pd
+import pytest
 
 import cudf
 from cudf.testing import assert_eq
@@ -14,3 +15,8 @@ def test_tz_convert():
     pidx = pidx.tz_localize("UTC")
     idx = idx.tz_localize("UTC")
     assert_eq(pidx.tz_convert(tz), idx.tz_convert(tz))
+
+
+def test_tz_convert_naive_typeerror():
+    with pytest.raises(TypeError):
+        cudf.date_range("2020", periods=2, freq="D").tz_convert(None)
