--- python/cudf/cudf/tests/indexes/test_categorical.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
