--- python/cudf/cudf/tests/indexes/test_conversion.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
