--- python/cudf/cudf/tests/indexes/test_multiindex_compat.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
