--- python/cudf/cudf/tests/indexes/test_sorting.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
