--- python/cudf/cudf/tests/indexes/test_time_specific.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
