--- python/cudf/cudf/tests/lists/test_list_methods.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
