--- python/cudf/cudf/tests/strings/test_string_methods.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
